requirements.txt
streamlit==1.37.1
pandas==2.2.2
altair==5.3.0
gspread==5.12.0
google-auth==2.23.4
